move max_procs def to utils/constants
