+++ src/cleanvision/issue_managers/duplicate_issue_manager.py
@@ -10,7 +10,8 @@ from tqdm import tqdm
 from cleanvision.issue_managers import register_issue_manager, IssueType
 from cleanvision.utils.base_issue_manager import IssueManager
 from cleanvision.utils.constants import SETS, DUPLICATE
-from cleanvision.utils.utils import get_max_n_jobs, get_is_issue_colname, MAX_PROCS
+from cleanvision.utils.utils import get_max_n_jobs, get_is_issue_colname
+from cleanvision.utils.constants import MAX_PROCS
 
 
 def get_hash(image: Image, params: Dict[str, Any]) -> str:
